feat: create test for checking pg and first round cutoffs

+++ src/main/kotlin/edu/mit/reserve/models/Category.kt
@@ -2,5 +2,5 @@ package edu.mit.reserve.models
 
 class Category(
 	val name: String,
-	val odds: Float
+	val odds: Double
 )
